Add files via upload
this is an ongoing project. and this little tool should not be used to disturb others. Just and only educational tool. To be able to determine where you received a call for which you do not know the country code. Easy to use with no need to connect to the internet